[CLANG] Fix ARM64 warning cleanup issues
Add missing ARM64 intrinsic declarations and Clang warning gates needed by the build.

Tighten signedness, enum, format, and narrowing conversions reported by Clang across affected user-mode, kernel, WDF, filesystem, and loader code.

Fix the NDIS 6 legacy packet wrapper to use the valid Reserved[0]/Reserved[1] slots instead of indexing past the two-entry Reserved array.

+++ dll/opengl/mesa/get.c
@@ -625,7 +625,7 @@ void gl_GetBooleanv( GLcontext *ctx, GLenum pname, GLboolean *params )
 	 *params = FLOAT_TO_BOOL(ctx->Point.Size );
 	 break;
       case GL_POINT_SIZE_GRANULARITY:
-	 *params = FLOAT_TO_BOOL(POINT_SIZE_GRANULARITY );
+	 *params = FLOAT_TO_BOOL((GLfloat) POINT_SIZE_GRANULARITY );
 	 break;
       case GL_POINT_SIZE_RANGE:
 	 params[0] = FLOAT_TO_BOOL(MIN_POINT_SIZE );
